+++ CHANGELOG.md
@@ -2,6 +2,10 @@
  
 ## [Unreleased]
 
+-
+
+## [0.3.0] - 2018-03-16
+
 - Asynchronous motor functions
 
 ## [0.2.0] - 2018-03-12